# contracts/api_contract.yaml
openapi: 3.1.0
info:
  title: Zygo Platform API
  description: |
    Complete API specification for the Zygo AI-driven workflow platform.

    ## Authentication
    All endpoints require authentication via JWT bearer token unless marked otherwise.

    ## Multi-Tenancy
    - **Tenant Mode**: Token includes `tenant_id` and `tenant_slug` claims
    - **Global Admin Mode**: Token includes `is_global_admin: true` claim

    ## Headers
    - `Authorization: Bearer {token}` - Required for authenticated endpoints
    - `X-Zygo-Mode: tenant | global-admin` - Mode indicator (informational)
    - `X-Zygo-Tenant-Slug: {slug}` - Tenant slug (informational, in tenant mode)
  version: 2.0.0
  contact:
    name: Zygo API Team
    email: api@zygo.tech
  license:
    name: Proprietary
    url: https://zygo.tech/terms
servers:
  - url: https://api.zygo.tech/v1
    description: Production API
  - url: https://api.staging.zygo.tech/v1
    description: Staging API
  - url: http://localhost:3001/v1
    description: Local development
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token with tenant_id and user_id claims

    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie for web clients
  schemas:
...
    TenantConfig:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [active, suspended, trial]
        plan:
          type: object
          properties:
            name:
              type: string
            tier:
              type: string
              enum: [free, starter, professional, enterprise]
            features:
              type: array
              items:
                type: string
        branding:
          type: object
          properties:
            logo:
              type: string
              format: uri
            primaryColor:
              type: string
              pattern: '^#[0-9A-Fa-f]{6}$'
            customDomain:
              type: string
        settings:
          type: object
          properties:
            ssoEnabled:
              type: boolean
            mfaRequired:
              type: boolean
            ipWhitelist:
              type: array
              items:
                type: string
        limits:
          type: object
          properties:
            maxUsers:
              type: integer
            maxNodes:
              type: integer
            maxExecutionsPerMonth:
              type: integer
            maxStorage:
              type: integer
              description: Storage limit in bytes
        usage:
          type: object
          properties:
            users:
              type: integer
            nodes:
              type: integer
            executionsThisMonth:
              type: integer
            storageUsed:
              type: integer
...
paths:
...
  /tenants/{tenantSlug}/config:
    get:
      tags: [Tenant]
      summary: Get tenant configuration
      description: Load tenant configuration including plan, features, limits, and usage
      operationId: getTenantConfig
      parameters:
        - name: tenantSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Tenant configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TenantConfig'
        '404':
          description: Tenant not found
        '403':
          description: Access denied
...